Refactor DetectionRulesClient
