+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_management/logic/detection_rules_client/__mocks__/detection_rules_client.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { IDetectionRulesClient } from '../detection_rules_client';
+import type { IDetectionRulesClient } from '../detection_rules_client_interface';
 
 export type DetectionRulesClientMock = jest.Mocked<IDetectionRulesClient>;
