+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_monitoring/logic/detection_engine_health/rule_objects/fetch_rule_by_id.ts
@@ -11,7 +11,7 @@ import type {
   RuleObjectId,
   RuleResponse,
 } from '../../../../../../../common/api/detection_engine/model/rule_schema';
-import { readRules } from '../../../../rule_management/logic/rule_management/read_rules';
+import { readRules } from '../../../../rule_management/logic/detection_rules_client/read_rules';
 import { transform } from '../../../../rule_management/utils/utils';
 
 // TODO: https://github.com/elastic/kibana/issues/125642 Move to rule_management into a DetectionRulesClient
